[#3134] comparison of integers of different signs

+++ plugins/database/src/general_query.cpp
@@ -301,7 +301,7 @@ tScan( int table, int link ) {
     }
 
     thisKeep = 0;
-    if ( table < 0 || table >= sizeof( Tables ) / sizeof( *Tables ) ) {
+    if ( table < 0 || static_cast<std::size_t>(table) >= sizeof( Tables ) / sizeof( *Tables ) ) {
         printf( "index %d out of bounds.", table );
         return -1;
     }
